move one_times to ufl_utils.py
